Add metadata and sitemap
