+++ app/form/page.tsx
@@ -1,4 +1,10 @@
 import { type FC } from 'react'
+import type { Metadata } from 'next'
+
+export const metadata: Metadata = {
+  title: 'Form - Dane Willacker',
+  description: 'General purpose form to connect with Dane Willacker.',
+}
 
 const FormPage: FC = () => {
   return (
