+++ public/sitemap.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <url><loc>https://danejw.com/</loc></url>
+  <url><loc>https://danejw.com/about</loc></url>
+  <url><loc>https://danejw.com/contact</loc></url>
+  <url><loc>https://danejw.com/hire-me</loc></url>
+  <url><loc>https://danejw.com/customer-feedback</loc></url>
+  <url><loc>https://danejw.com/form</loc></url>
+  <url><loc>https://danejw.com/newsletter-signup</loc></url>
+  <url><loc>https://danejw.com/services</loc></url>
+  <url><loc>https://danejw.com/services/knowledge-base</loc></url>
+  <url><loc>https://danejw.com/services/web-app-development</loc></url>
+  <url><loc>https://danejw.com/services/webapp/estimator</loc></url>
+</urlset>
